apps/dash-board/pyproject.toml
[tool.poetry]
name = "dash-board"
version = "0.1.0"
description = ""
authors = ["Timon Borter <timon.borter@gmx.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dash = "^2.16.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pandas = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
